src/emergency_stop.py
from pubsub import pub

def check_safety():
    # Check for emergency stop conditions
    pass

def main():
    # Subscribe to relevant topics
    pub.subscribe(check_safety, 'safety_data')

    # Main loop
    while True:
        pass

if __name__ == "__main__":
    main()
