src/path_finding.py
from pubsub import pub
import dijkstra_algorithm

def process_lidar():
    # Process lidar data to create a graph for dijkstra's algorithm
    pass

def plan_path():
    # Use dijkstra's algorithm to plan a path
    pass

def main():
    # Subscribe to relevant topics
    pub.subscribe(process_lidar, 'lidar_data')

    # Main loop
    while True:
        plan_path()

if __name__ == "__main__":
    main()
