requirements.txt
Flask==2.1.1
pypubsub==4.0.3
pyserial==3.5